Backend/requirements.txt
apns2==0.4.1
asn1crypto==0.24.0
Babel==2.6.0
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
Django==2.1.2
django-braces==1.13.0
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-phonenumber-field==2.1.0
django-push-notifications==1.6.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
exponent-server-sdk==0.2.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.7
Markdown==3.0.1
oauthlib==3.0.1
phonenumbers==8.10.4
Pillow==5.4.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py-vapid==1.5.0
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.5
pywebpush==1.9.3
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
urllib3==1.24
